20questions/requirements.txt
streamlit
pandas
google-auth
google-auth-oauthlib
google-auth-httplib2
google-auth-service-account
gspread
python-dotenv
